+++ sus/libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: otchekai <otchekai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 13:10:27 by otchekai          #+#    #+#             */
-/*   Updated: 2022/10/16 06:49:34 by otchekai         ###   ########.fr       */
+/*   Updated: 2022/10/17 15:19:32 by otchekai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int				i;
+	size_t				i;
 	unsigned char	*str;
 
 	str = (unsigned char *)b;
